fix: Rename `init_data` to `data` in config flow
Seems that `init_data` might have been kind of a reserved variable,
and since we used it, HomeAssistant attempted to do some cleanup when
finishing the config flow. Which threw an error, and might be the cause
of #27 as well. I'm unable to test properly, since I don't have access
to more than one device.